+++ crates/ruma-client-api/CHANGELOG.md
@@ -9,6 +9,7 @@ Breaking changes:
 Improvements:
 
 * Add unstable support for refresh tokens (MSC2918)
+* Add `ErrorKind::{UnableToAuthorizeJoin, UnableToGrantJoin}` encountered for restricted rooms
 
 # 0.14.1